Refactor and fix TypeScript (again)
