+++ packages/core/test/collaboration/mergeNode/onLegacyEmptyTextNode.tsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { Editor, Transforms } from 'slate';
 import { jsx } from '../../support/jsx';
-import { yTextFactory } from '../../yTextFactory';
+import { yTextFactory } from '../../support/utils';
 
 export const input = (
   <editor>
