+++ packages/core/tsconfig.build.json
@@ -1,8 +1,8 @@
 {
   "extends": "../../config/typescript/tsconfig.json",
-  "include": ["src/**/*", "test/**/*"],
+  "include": ["src/**/*"],
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "./src",
     "outDir": "./dist",
     "declarationDir": "./dist",
     "composite": true
