+++ packages/react/tsconfig.json
@@ -7,5 +7,5 @@
     "composite": true,
     "jsx": "react"
   },
-  "references": [{ "path": "../core" }]
+  "references": [{ "path": "../core/tsconfig.build.json" }]
 }
